ASoC: firmware: Add support for dynamic components
Add generic support to create ASoC kcontrols, widgets, graphs and coefficients from
firmware files.  This is useful for many modern audio DSP devices where the firmware
often defines the graph and mixers and allows the vendor firmware file to ship with the
ASoC specific components along with the firmware text.

The generic FW format is very simple and uses a small header to describe each file
section. e.g. a FW file with text and kcontrols would have a header for the text and
another header for the kcontrols section. The sections can be in any order within
the firmware file meaning many different FW section files can be cat'ed together into
a larger file.

This patch creates the generic ASoC firmware core and an associated header API
that can also be used by userspace firmware generation tools. It also defines a
small API that can be used by ASoc component drivers to interrogate and process
the firmware data. Most firmware data is handled by the core, e.g. widgets, kcontrols
and graph, whilst other data is passed directly to the component driver for
enumeration e.g. text.

This patch has working support to load ASoC :-

 o Kcontrols
 o TLV Kcontrols
 o DAPM widgets
 o DAPM Graph
 o Simple Enum Kcontrols with coefficients.
 o Firmware Text
 o Custom vendor blocks

There is a command line firmware file generation tool that is mostly complete at :-

git@gitorious.org:omap-audio/asoc-fw.git

This tool generates the headers and generic components based on the ASoC types above
and it's intended to be upstreamed in to alsa-utils soon. There is example code with
the tool for generating firmware for the OMAP4 ABE.

Change-Id: I506e7c39f4c2a8c0fd2d51a3dae487a7baf588cf