audio: Fix for Codec power off after the first playback during the multistream playback.
BZ: 132116

Codec is getting powered off, when the capture stream is still active.

The codec should not be powered off, when the capture stream is active.
During multi stream playback, as soon as the first playback ends, the
snd_pcm_close() will be called and since playback counter is still active,
the codec gets powered off even when the capture stream is active, due to a
wrong condition check in the soc core.

This patch does not power off the codec when the playback or capture streams
are still active.

For K3.4:
	BZ: 49949.

Change-Id: I6954e3dc307e64fdf3f7a6e22d17ec162ce05025
Orig-Change-Id: I81436b9b5425c6733fed997b3b5e5a8dd355cae0
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Signed-off-by: Praveen Diwakar <praveen.diwakar@intel.com>
Reviewed-on: http://android.intel.com:8080/123316
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-on: https://android.intel.com/217111
Tested-by: Koul, Vinod <vinod.koul@intel.com>