ALSA:[MOVE] hda - Create common HD Audio Codec Library
BZ: NONE

This patch creates an HD Audio Codec Library.
New hda directory is created under sound so that this library
can be used by both legacy ALSA HDA driver and ASOC HDA codec
drivers.

NOTE: This uses COPY to code as we are not modfying existing code to use the
libraries.No code changes done which should be done eventaully for upstreaming
so there it would be code move ONLY

MOVE FILES=>
 {sound/pci/hda => include/sound}/hda_auto_parser.h
 {sound/pci/hda => include/sound}/hda_beep.h
 {sound/pci/hda => include/sound}/hda_codec.h
 {sound/pci/hda => include/sound}/hda_generic.h
 {sound/pci/hda => include/sound}/hda_jack.h
 sound/{pci => }/hda/hda_auto_parser.c
 sound/{pci => }/hda/hda_beep.c
 sound/{pci => }/hda/hda_codec.c
 sound/{pci => }/hda/hda_eld.c
 sound/{pci => }/hda/hda_generic.c
 sound/{pci => }/hda/hda_hwdep.c
 sound/{pci => }/hda/hda_intel_trace.h
 sound/{pci => }/hda/hda_jack.c
 sound/{pci => }/hda/hda_proc.c
 sound/{pci => }/hda/hda_sysfs.c
 sound/{pci => }/hda/hda_trace.h

Change-Id: Ibdc09dd9e6392be415f6eb0dba4f84b41a789018
Signed-off-by: Lakshmi N Vinnakota <lakshmi.n.vinnakota@intel.com>
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Reviewed-on: https://android.intel.com/210266