Audio:hda set the correct stream format in codec
BZ:NONE

cvt format will be updated when the stream format is
updated in codec . Format will be updated if the current
stream format is different from the previous format(cvt format).
In the hw_params, the cvt format was set to the current stream
format, becase of this the cvt fmt and stream fmt will be
same and the stream format will never be set in the codec.
The codec fmt will always be set to default value

This patch fixes the above issue by setting the stream format
in dma params in hw_parms and the same will be passed in update
_pcm_format instead of cvt format.

Change-Id: I3b5421c590ff35ea989cf037872a0a6ab93fa465
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Reviewed-on: https://android.intel.com/212700