ALSA: pcm: Uninline snd_pcm_stream_lock() and _unlock()
The previous commit for the non-atomic PCM ops added more codes to
snd_pcm_stream_lock() and its variants.  Since they are inlined
functions, it resulted in a significant code size bloat.  For reducing
the size bloat, this patch changes the inline functions to the normal
function calls.  The export of rwlock and rwsem are removed as well,
since they are referred only in pcm_native.c now.

Change-Id: I34c1af247eb83973c58c2b2d354019878e234993
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Reviewed-on: https://android.intel.com/278405
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Kp, Jeeja <jeeja.kp@intel.com>