Set proper micbiases for amic and dmic path
BZ: 0

Micbias needs to be on when amic or dmic turning on is required
For DMIC also micbias is required to be on, so that codec can
generate DMIC clock. Add micbiases as a part of DAPM route
based on WM8281 AOB for MOFD. Need to change the route if there
is any change in AOB schematic later.

Remove some mis-leading comments as well.

Change-Id: Ibbd23ccc49919e022a0d278d7745146a5351fddf
Signed-off-by: Hardik T Shah <hardik.t.shah@intel.com>
Reviewed-on: https://android.intel.com/279750
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Kp, Jeeja <jeeja.kp@intel.com>