Audio: configure DMIC mode selection
BZ:NONE

This patch adds DMIC mode selection. for DMIC23,
DMIC23 will be used for 16k and will be set to
speech mode. 1 means speech mode and 0 means
high quality mode(48k).
currently we are not sure of the DMIC node structure.
after refactoring of DMIC structure in FW, this need
to fixed.

Change-Id: Iba37a9194f16af8cc277eb35d2fed4cc61fbb1b4
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Reviewed-on: https://android.intel.com/285568