Audio: fixed format valid bit depth based on input param
BZ:NONE

if the input stream params is 16 bit, then we need to set the
bit depth to 16 bit container and in case of 24 bit
set the bit depth to 32 bit container.

this patch overwrites the module default value. we need to fix
it without overwriting the default. so putting this as workaround.

Change-Id: I7cffdc0246385baaf057d5f0562807a096ab240c
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Reviewed-on: https://android.intel.com/285569