Audio: Intel: HDA: correct bdl_pos_adj
HSD: 46247

From HD Audio spec Section 3.6.2 Buffer Descriptor List,
BDL must be 128 byte aligned, if not it can have undesirable
issues. Currently bdl_pos_adj when set to 1 was resulting
in the BDL pos_adj to 4 bytes. As a result of which the first
BDL would be 4 bytes and the next BDL address would be 4byte
aligned instead of 128 byte aligned. So the bdl_pos_adj is now
set to 0 which ensures 128 byte alignment. This helped to
resolve PCM Error scenario during playback.

Change-Id: I3a1a75f7873a4ee11a1f2e2988c20b64e078c312
Signed-off-by: Mythri P K <mythri.p.k@intel.com>
Reviewed-on: https://android.intel.com/304889
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>