ASoC: hda: intel: Fix to check intermediate power state
With the changes in the dapm framework for the widget power state, power
update is delayed until changes are actually written to the hardware. Refer
commit: "ASoC: dapm: Delay w->power update until the changes are written"

new_power member of the widget holds the intermediate power state before the
power state is updated with actual value.

This doesn't hold true to check on the power state during pre_pmd/post_pmd
for bind/unbind. So check on the intermediate new_power state for
bind/unbind.

Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>