Revertme: ASoC: hda: Start pipeline after inter pipe modules are bound
BZ:0

With addition of more than 2 modules in pipeline, pipe was
getting started even if the intermodule bind has not happened.
Make sure that inter module bind has happend and then start
the pipeline.

This patch will be reverted once the correct sequence of
module initialization and binding is implemented

Change-Id: I6b87874a2ebd7754b02f899630e4f73debeaf966
Signed-off-by: Hardik T Shah <hardik.t.shah@intel.com>
Reviewed-on: https://android.intel.com/313952
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Kp, Jeeja <jeeja.kp@intel.com>