ASoC: Intel: Skylake: Fix CLDMA buffer copy when ring buffer wrap occurs
When downloading the firmware/module, if ring buffer boundary is
reached, need to wrap to the zeroth position. When copying the
buffer, need to copy till end of buffer and the remaining buffer
needs to be copied from zeroth position. Copy was not handled
correctly in case when wrap condition is reached which caused
invalid data to be copied resulting in invalid module hash failure.
This patch fixes the issue by handling copy at the boundary
condition correctly.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>