ASoC: Intel: Skylake: Add check for format FIXUP
When format FIXUP is added, in case of 16 bit sample size
bit depth and valid_bit_depth are made equal to 16. This
Change is effecting when other fixup's are added.
eg. rate and channel.
This patch will check for format fixup. if it is, then
only it will make both bit depth and valid_bit_depth to 16 bit.

Change-Id: I3a624f6e7b2e3ff3077ca4068be556198c2f69b8
Signed-off-by: Kranthi G <gudishax.kranthikumar@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/7533
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>