ASoC: Intel: Skylake: Override Copier DMA FIFO size
The DMA FIFO sizes should ideally be taken from the
topology binary. Instead, Override the copier DMA FIFO
size to correspond to 2ms for normal streams and 50ms
for deepbuffer streams.

Also provide a module parameter to tune the FIFO size of
deep buffer streams. This tuning will be done for D0i3
(and hence S0ix) residency and then the value will be
set in topology binary.

Change-Id: I324c498886e42bd597a31fdfec5f15a38780c253
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/7982
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>