drm/i915: enable IPC when i915.enable_ipc=1 (v3)
Enabling IPC (Isochronous Priority Control) is
recommended as a fix for display corruption and
flickering issues found on APL-I Silicons.

Until this becomes our offical fix for issues,
we keep this as an optional configuration
that can be conditionally enabled.

To make it enabled, i915.enable_ipc=1 should be
provided as one of boot command parameters.
It is turned off by default.

v2: fixed syntax
v3: IPC can only be enabled on BXT

Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>