ASoC: core: Introduce DAPM sink
DAPM models various widgets but lacks a sink widget.
DSPs can have modules which take audio data, process it
and are capable of generating events thus acting as a
sink of data.

To make the dapm graph complete for such paths we need a
dapm sink widget for these modules, so add a
SND_SOC_DAPM_SINK to declare such a widget.

Change-Id: I200c490c3cb2ea841a8ac6d23d5b35bd11812967
Signed-off-by: Divya Prakash <divya1.prakash@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/8108
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>