ASoC: Intel: Boards: Add control in IVI Machine driver for Sample Rate
kcontrol has been added to accept the sample rate as a parameter
from user, for BT usecases. Supported rates are 8K and 16K, which
are set before start of a usecase. This is needed for just one SSP
for both PB and CP operations.
Based on this rate input, the modules in the device pipe are
configured using appropriate fixups.

Change-Id: I6109840e0f493ce08f19348d1a0e3c1f08274063
Signed-off-by: Mousumi Jana <mousumix.jana@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/8080
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>