drm/i915/modeset: parse video= command line parameter for connector (v2)
and bypass the query on all connectors.  If no video= parameters are
present or none are valid, fall back to querying each connector for
its status. If there is one or more valid video= parameters, only
the specified connectors will be configured.

This only looks for the force option in the video= options. For example:

video=DP-2:d   - Disable this connector (ignore it)
video=DP-1:e   - Enable this connector (force)

v2: cleaned up the backoff on deadlock handling (squash to previous) (Bob)

Signed-off-by: Bob Paauwe <bob.j.paauwe@intel.com>
Reviewed-by: James Xiong <james.xiong@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>