ALSA: core: add .update_appl_ptr callback for pcm ops
When appl_ptr is updated let low-level driver know, e.g.
to let the low-level driver/hardware pre-fetch data
opportunistically.

The existing .ack callback could be used but it would
need to be extended with new arguments, resulting in
multiple changes in legacy code.

Instead a new .appl_ptr_update callback is added.
The difference between .ack and .appl_ptr_update is that
.ack is only called on read or write. .appl_ptr_update
is called on read, write, rewind, forward or when updating
the appl_ptr from userspace.

In the ALSA core, this capability is independent from the
NO_REWIND hardware flag. The low-level driver may however
tie both options and only use the updated appl_ptr when
rewinds are disabled due to hardware limitations.

Change-Id: I8e01bdfb0d9472829e2a371e4faecfd59b3020f6
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/9562
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>