ASoC: Intel: Skylake: Fix ipc error during simultaneous extractor and injector
In an simultaneous extractor-injector usecase, the injector probe point
was getting set before the stream is opened and before any DMA is attached
which should not be the case. This patch ensures that the injector probe
point is not set unless a new compr playback stream is opened and
a DMA is attached.

This patch also resolves the issue with probe disconnect.

Change-Id: Ifef3c5e33ab48b106feb3d6066444d9ba00cc00d
Signed-off-by: Pawse, GuruprasadX <guruprasadx.pawse@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/9395
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Babu, Ramesh <ramesh.babu@intel.com>