ASoC: Intel: Skylake: Add strip extended manifest utility
Some upcoming platforms like broxton etc have extended manifest
in firmware binary. This is not required to be downloaded to DSP.
So driver needs to strip this before downloading.

Add a utility function to check if a header exists, and remove it
in that case

Change-Id: I8cf36cb3787f34cc7f293997fd3dbb085e0099df
Signed-off-by: Ramesh Babu <ramesh.babu@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/11812
Tested-by: Sangaraju, KarthikeyaX <karthikeyax.sangaraju@intel.com>
(cherry picked from commit df6abe83b93291d624af7d707e456c56bac8e58b)