drm/i915: Add aux plane verification in addFB2
A recent change in the upstream kernel failed a call to addFB2 when
for n planes (derived from the pixel format), the n+1 structure for
pitches/offsets/handles/modifier in the framebuffer, contained non-zero
values. This extra data being considered as garbage.

Render compression is supported on planes with pixel format RGB8888,
so number of planes is considered as 1. Aux stride and offset values
are passed in the next offset of the pitches/offset array by userspace.
By the existing logic, this would be treated as garbage..

To avoid this incorrect failure in case of render compression, this
patch introduces a check related to AUX plane to support the scenario
of render compression. When AUX_PLANE flag is set, num_planes is
forcibly incremented so that the following for loop does not fail on
account of garbage data.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7860
Signed-off-by: Vandana Kannan <vandana.kannan@intel.com>
Reviewed-on: https://android.intel.com:443/446478
Signed-off-by: Jeff McGee <jeff.mcgee@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>