ASoC: Intel: Skylake: Add support for topology change notification
Some events like Pipeline start, pipeline deletion, DSP D0/D3
need to be notified to the user, in order to convey a
change in the topology. Support for notifying such events has
been added by the way of a kcontrol. This kcontrol has the
time at which the last change occurred in the topology.

Change-Id: I9b0d4fbf48e3ba614c52aece12b33810c8a4ec56
Signed-off-by: Mousumi Jana <mousumix.jana@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/10167
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>