ASoC: Intel: Skylake: Exit gracefully on base FW/library load failure
FW/Library load failure was causing kernel crashes. Inorder to handle
such failures exit and gracefully, the following changes are done.

1. Added a flag "is_first_boot" to specify if the first boot after power
up is yet to be done. Before first boot completes or on failure of first
boot due to FW/Library load issues, this flag will remain true.

2. If first boot is yet to be done(i.e. flag = true)  do not call
the DSP suspend/resume functions.  But the controller suspend/resume
will still work.

3. On failure of soc probe due to FW/lib load failure or otherwise,
set the platform pointer of skl_sst  to NULL. This is because, on
soc probe failure the memory for soc_platform stucture will be freed,
hence the copy of this pointer in skl_sst should be set to NULL to
avoid improper access during controller suspend.

Change-Id: Id287d5bdf5a21f1c678d57342ccf135ec1163a49
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/10553
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>
Tested-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>