ASoC: Intel: Set all I2S ports to slave mode after DSP power up in BXTP
During all DSP HW power up sequences, the I2S ports are defaulted
to Master mode for Frame sync and bit clock. Due to this, the
pins drive a higher voltage and cause issues on some boards.
To prevent this, the ports have to be brought into slave mode,
before the ROM execution.

Change-Id: I6668f75d4c5701f49afebbf96947db9f0adf7f4d
Signed-off-by: Mousumi Jana <mousumix.jana@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/10524
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Tested-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>