ASoC: Intel: Skylake: Probe - Fix DMA assignment and reset
1. Extractor DMA is to be assigned when the first probe stream(irrespective
of whether it is injector or extractor) is opened. But if the first probe
stream is injector, we get injector's substream pointer and we do not
have the right substream pointer for extractor.

The existing code passed injector's substream while assigning extractor
DMA.This patch sets NULL for substream while assigning the DMA for
extractor and sets the correct substream pointer later when open is
indeed for extractor.

2. DMA reset for exractor should not be done after  probe module is
initialized.

The existing code reset DMA in hw_params. This can result in DMA reset
after probe module init when  injector and extractor are started one
after the other.

This patch moves the DMA reset from hw params to immediately after DMA
assignment in open call back for both injector and extractor.

Change-Id: I8bc37b55af3068aed5f3b1547ec2fb0e19e4565d
Signed-off-by: Pawse, GuruprasadX <guruprasadx.pawse@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/10830
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Kale, Sanyog R <sanyog.r.kale@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>