ASoC: Intel: Skylake: Update module id in pin connections
Currently, each module instance id is getting read from
firmware binary and updated in the driver. However, the
module id in the statically mapped pins of a module is
still the same as coming from topology binary. This causes
module binding failure if the module id is wrong in the
topology binary. This patch updates the module id in the
pin connections of each module instance.

Change-Id: I1bdec44601b660a6c6ff602d90e0832b2516f825
Signed-off-by: Panwar, Ashish <ashish.panwar@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/11099
Reviewed-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>
Tested-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>