ASoC: Intel: Skylake: Decouple/Recouple for Stream Format
Due to a bug in DMA FIFO HW, the controller needs to be put in
coupled mode before performing writes to the Stream Format Register.
Once the write is completed, it can be reconfigured back to
de-coupled mode
This WA is needed to avoid roll over of the DMA Buffer in the FW
code, before even the DSP consumes data. If not done, this can result
in audio being played slowly as the DSP inserts silence data to
complete 1ms frame.

Change-Id: I439b4d13e2a18aa1b90cf959a251835479f90903
Signed-off-by: Pardha Saradhi K <pardha.saradhi.kesapragada@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/11762
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Sangaraju, KarthikeyaX <karthikeyax.sangaraju@intel.com>