drm/i915/gen9: Implement turbo enable/disable for GEN9
Update ips_turbo_disable() entry point to work on GEN9. When turbo is
disabled we don't allow frequencies above RP1 to be set.

Add a new entry point that can be used to enable GPU turbo. When turbo
is enabled, the driver will only limit frequencies based on the softmax
value.  This is the default behavior.

Signed-off-by: Bob Paauwe <bob.j.paauwe@intel.com>
Reviewed-by: Kai Chen <kai.chen@intel.com>
Reviewed-by: Olga Koteoglou <olga.koteoglou@intel.com>