ASoC: Intel: Skylake: Correct the Probe buffer size
Probe module's buffer size is hard coded in the driver to
a value that is topology dependant. Special topologies
like ULL which operate on optimized memory spaces, cannot
afford a large value of the Probe module's input buffer.
This patch provides a work around to the problem by
defining a small value specific to the ULL topology.

Change-Id: I3977d2ab671505991a3fdeb674b5cdd7d899c7c8
Signed-off-by: Pardha Saradhi K <pardha.saradhi.kesapragada@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/12163
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Sangaraju, KarthikeyaX <karthikeyax.sangaraju@intel.com>