ASoC: Intel: Skylake: Support for Probe DMA Buffer size
This patch removes hard coding of buffer size for Probe
Module. For probe, DMA Buffer size is calculated as-

2 * [Probe Hdr + ((SRate/1000)*(ch)*(sch_time_ms)*(bytes per sample)) +
Probe Footer]
where, Probe Header and footer are 24 and 8 bytes respectively.

Hence this value is topology specific and is to be used
from XML as dma_buffer_size value for Probe Module.

Change-Id: I060f265a6dfe51fb1dc3abcdc75975322372347b
Signed-off-by: Mousumi Jana <mousumix.jana@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/12614
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Sangaraju, KarthikeyaX <karthikeyax.sangaraju@intel.com>