drm/i915/dsi: Set/Clear panel control GPIO.
APL platforms have a GPIO pin that enables panel control. This pin
must be set after the power on GPIO pin and before the backlight
enable GPIO pin.  However, none of the existing MIPI sequences do
this.  Also, the pin does not seem to be defined in the ACPI table
so it is not possible to use the consumer GPIO api to control this
pin.

Add the necessary defines and functions based on the older gpio
functions and use those to set/clear this pin during prepare/unprepare.

Signed-off-by: Bob Paauwe <bob.j.paauwe@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>