drm/i915/bxt: DSI dual-link shift WA for A0 (v2)
When using dual-link DSI, there can be a split screen display
corruption if the planes are enabled without a pipe scaler
being enabled.

This always enables a pipe scaler if on A stepping hardware. This
is turning on the scaler even in cases where the WA isn't needed.

This should only be needed for A0 hardware and is untested.

Based on a similar patch:
	Author: Vidya Srinivas <vidya.srinivas@intel.com>
	Date:   Fri Nov 25 13:48:59 2016 +0530

	Scalar fix (split)

v2: detect wa needed in atomic check and setup scaler there. (Matt)
    remove accidental '{}'s (Matt)
    use IS_BXT_REVID() macro (Matt)

Signed-off-by: Bob Paauwe <bob.j.paauwe@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>