drm/i915/gen9+: Add auxiliary surface register definitions
In addition to the regular surface offset registers, gen9+ platforms
also have registers for an auxiliary surface.  When displaying planar
YUV formats like NV12, the Y data is programmed as the main surface
while the UV data is programmed as the auxiliary surface.  The auxiliary
surface is also used when the display controller is expected to do
render decompression; in this case the auxiliary surface contains the
location of the color control surface (CCS).

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Bob Paauwe <bob.j.paauwe@intel.com>