ASOC: Intel: BXT: Resume time optimization
During S3->S0 transition, sometimes ROM init fails because of
authentication engine loads later than the OS.
In this case driver waits for a longer period and then
retries the FW download causing huge delay in resume time of
audio device.

To avoid this, ROM INIT wait time is set to a optimal value and
increased the retries for firmware download.

Change-Id: I83c8847e5dba921951d52e38713afd5c64414735
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: hsvijay <vijay.h.s@intel.com>