drm/i915: Don't always treat hung GPU as wedged
While creating our forklift, we lost some important logic during the
reset handling.  Upstream checks the request's fence error code to
determine whether a reset failed; if so, it marks the GPU as wedged.
Our base kernel doesn't have dma_fence error codes, so we don't have
this information to check.  Just assume the reset was successful for now
(which is better than our previous situation of assuming it always
failed, leading to every reset resulting in a wedged GPU).

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>