drm/i915: Program scaler phase for NV12
Unfortunately the bspec doesn't really make clear what these registers
do or why we need to program them.  Based on VPG's Android code, it
appears leaving them as 0 is fine for packed formats, but they need to
be set to other values for NV12.  The VPG code indicates that they could
be tuned by userspace according to the actual buffer being used (but not
how that tuning would work), but that the values here are sensible
defaults.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Acknowledged-by: Bob Paauwe <bob.j.paauwe@intel.com>