drm/i915: Update watermark programming for NV12
The behavior here is a bit strange --- according to spec, we're supposed
to calculate watermark values for both the Y and the UV planes to ensure
that they fall within the relevant DDB allocations, but then we only
actually program the Y surface watermarks into the hardware registers.
So the fact that this patch calculates (and checks) UV values and then
immediately clobbers those values is expected based on the current
contents of the bspec.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Bob Paauwe <bob.j.paauwe@intel.com>