drm/i915: Restrict NV12 support to APL E0+ (v2)
Note that SoC stepping E0 uses D0 GT/Disp stepping, so the actual test
in the code is for 'D0' revision ID's.

This should technically be squashed into the previous patch, but I'm
supplying it separately to make it easier for validation to temporarily
drop while developing their test cases.

v2: Rebase on previous changes.  The test now checks up to C0 (the last
    GT/Display stepping before D0) for platforms that can't support
    NV12, rather than D0 and up for platforms that can.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Bob Paauwe <bob.j.paauwe@intel.com>