drm/i915: Do clflush before moving obj to GPU if cache is dirty
(IOTG only patch, not going to be upstreamed)
In i915_gem_execbuffer_move_to_gpu, we do clflush only if obj
is in I915_GEM_DOMAIN_CPU. This case does not cover when obj
is other domains while its cache_dirty flag is being set. We
need to perform clflush in this case as well.

Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>