drm/i915: Move cursor watermark writes back under vblank evasion
According to the bspec, it should be safe to write watermark registers outside
of vblank evasion since the watermark registers are double buffered and armed
by writes to the surface offset register.  Unfortunately, we've found that the
hardware seems to behave incorrectly when we have a sequence of:

  1) Write cursor watermarks
  2) Vblank happens
  3) Write cursor control register
  4) Write cursor base offset register

With this sequence, the write of the cursor control register doesn't arm.  This
is most problematic when the cursor is being disabled --- we wind up setting
the cursor base to 0x0, but not actually succeeding in disabling the cursor
plane.  This can lead to the cursor scanning out whatever garbage happens to be
pointed to by the GTT entries at offset 0, or it can lead to a storm of GTT
fault interrupts which can effectively DoS the system.

We'll need to investigate this problem further and provide evidence to the
hardware team, but in the meantime, let's work around the hardware behavior by
moving the cursor watermark updates back under vblank evasion.  This will ensure
that no vblank can occur between the watermarks and the control register write.

HSD: 1504434395 ("[LH] Corruption on DP/HDMI and all MIPI panels, WW10.3 Dev")
HSD: 1504468196 ("[LH] Firmwareless bios Fault Errors after killing weston")
Cc: Bob Paauwe <bob.j.paauwe@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Bob Paauwe <bob.j.paauwe@intel.com>