drm/i915: calculate aux_offset and aux_stride differently for RBC (v5)
Current method for calculating stride and offset for aux plane
is valid only when surface format natively requires multiple
planes. When RBC is enabled for single-plane surface format,
this method returns incorrect stride and offset for aux plane
because # of planes is still set as 1. So we are adding
a special path for this case.

v2: 1. aux_stride is now calculated only once
    2. same calculation will be done on update_plane
       for the sprite plane.
    3. pre-programmed aux_x and aux_y will be used
       if RBC is enabled.

v3: 1. added render_comp flag as input parameter to
       skl_plane_stride so that stride can be calculated
       when RBC is enabled
    2. defined and added skl_plane_aux_offset_rbc to
       get aux_offset for RBC case calculated
    3. cleaned up skylake_primary_update_plane and
       skl_update_plane to call functions added (or
       modified) in 1, 2 to calculate stride and offset

v4: 1. Don't need to reset bitmask,
       PLANE_CTL_DECOMPRESSION_ENABLE
       because it's not set anywhere before render_comp is
       checked. So removed it.
    2. removed # of planes check when
       render_comp is checked in skylake_update_primary_plane

v5: 1. pre-calculation of aux_offset in check_plane_surface
    2. minor fixes

Signed-off-by: Kim, Dongwon <dongwon.kim@intel.com>
Signed-off-by: Bob Paauwe <bob.j.paauwe@intel.com>
Reviewed-by: Bob Paauwe <bob.j.paauwe@intel.com>