ASoC: Intel: Skylake: Support multiple format configs
A module can have two set kinds of set params, as per
topology requirements. For example, one pre-init and one post-init.
But currently, there is support for just one type, as the
format_config.

This patch extends the format_configs to 4, so as to
be able to support pre-init, post-init and post-bind
type of set params, for the same module, simultaneously.

Change-Id: I46ef065be7f983ff602828b212c061eba7255e52
Signed-off-by: Pardha Saradhi K <pardha.saradhi.kesapragada@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/15542
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Sangaraju, KarthikeyaX <karthikeyax.sangaraju@intel.com>