drm/i915/guc: Hold the global timeline lock until we are sure the request is in the GuC's WQ
Otherwise another request could make it to the WQ before us, but after we have received a global
seqno (causing a seqno inversion).

This should not deadlock with the workqueue lock because that's only used to allocate other
requests (and grabbing wq_lock to check for free space in the wq is only best effort).

A possible improvement to this would be releasing the timeline lock as soon as we have placed
the request in the WQ, but before ringing the doorbell (since seqno ordering is already
guaranteed at that point).

Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>
Cc: Jeff McGee <jeff.mcgee@intel.com>
Signed-off-by: Jeff McGee <jeff.mcgee@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>