ASoC: tdf8532: Fix the codec status error issue on APL-GPMRB
Based on the TDF8532 manual:
If the wait_state result is ok, we should send CLK_DISCONNECT
command to force codec from STANDBY(2) to IDLE(1).
If the wait_state result is timeout, the codec state should be
at Clockfail(7), we still should send CLK_DISCONNECT command
force the codec from Clockfail(7) to Idle(1).

Signed-off-by: Wu Zhigang <zhigang.wu@linux.intel.com>
Reviewed-by: Keyon Jie <yang.jie@linux.intel.com>