[WORKAROUND] ASoC: SOF: start HDA DMA at hw_params() stage and remove stream_trigger()
It require HDA DMA to be available before trigger_start(), Here move
host side DMA start to earlier stage(hw_params) and remove
stream_trigger() as it is no need anymore.

This fix will help to remove HDA DMA workaround fix inside firmware.

TODO: revisit this patch for SOF 1.2, this removes the ability to use
the RUN bits for multi-stream synchronization and is likely not an
acceptable long-term solution

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>