ASoC: SOF: refine dapm route loading in sof
Some virtual routes and widgets may been added in topology to reusing
machine driver in Linux kernel. For virtual routes, both sink and source
are not buffer. Since only buffer linked to component is supported by
FW, others are reported as error, add check in route function, do not
send it to FW when both source and sink are not buffer

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>